chore: make BUILDS_YML_PATH not depend on the CWD
`path.resolve` uses the `process.cwd()` to resolve relative paths, which in turn makes this variable sensitive to the current working directory in which the node process is run.

This change makes it possible to run the build from any CWD. It's not a common thing to do, but I ran across it while trying to use some of these functions in a new build process. There might be more instances of CWD-sensitive paths, but this is the only one that I bumped in to.